+++ .gitignore
@@ -11,4 +11,7 @@ info.md
 /scraper/info.txt
 README.md
 PLANNETJE.md
-
+/data
+data
+/data/store.json
+/data/store
